third_party/blink/web_tests/platform/generic/external/wpt/html/webappapis/scripting/events/event-handler-processing-algorithm-error/document-synthetic-errorevent-expected.txt

This is a testharness.js-based test.
FAIL error event is normal (return true does not cancel; one arg) on Document, with a synthetic ErrorEvent assert_equals: expected (undefined) undefined but got (object) null
FAIL Initial values of ErrorEvent members assert_equals: expected (undefined) undefined but got (object) null
Harness: the test ran to completion.
